+++ MapbenderQueryBuilderBundle.php
@@ -2,10 +2,15 @@
 namespace Mapbender\QueryBuilderBundle;
 
 use Mapbender\DataSourceBundle\MapbenderDataSourceBundle;
+
 use Symfony\Component\Config\FileLocator;
+use Symfony\Component\Config\Resource\FileResource;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Extension\ExtensionInterface;
 use Symfony\Component\DependencyInjection\Loader\XmlFileLoader;
+use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
+
 
 /**
  * Mapbender Query Builder Bundle.